submit F.cpp to 'F - Takahashi's Basics in Education and Learning' (abc129) [C++14 (GCC 5.4.1)]

+++ 2019/0611_ABC129/F.cpp
@@ -328,7 +328,7 @@ Matrix<mint> affine(ll k)
   return X;
 }
 
-int f(int i)
+int f(ll i)
 {
   return (to_string(A + B * i)).size();
 }
